requirements.txt
numpy
pandas
geopandas
rasterio
shapely
matplotlib
seaborn
scikit-learn
torch
torchvision
torchmetrics
tqdm
pillow
folium
geemap
reportlab
scipy

